+++ asf_search/CMR/datasets.py
@@ -1,5 +1,13 @@
 from typing import List
 
+from asf_search.constants import PRODUCT_TYPE
+
+NISAR_PRODUCT_TYPES = [
+    PRODUCT_TYPE.RRSD, #L0
+    PRODUCT_TYPE.SME2, # L3
+    PRODUCT_TYPE.GSLC, PRODUCT_TYPE.GCOV, PRODUCT_TYPE.GUNW, PRODUCT_TYPE.GOFF, # L2
+    PRODUCT_TYPE.RSLC, PRODUCT_TYPE.RIFG, PRODUCT_TYPE.RUNW, PRODUCT_TYPE.ROFF,  # L1
+]
 
 dataset_collections = {
     'ALOS-2': {
@@ -54,21 +62,6 @@ dataset_collections = {
             'C1256358262-ASFDEV',
             'C1257349115-ASF',
             'C3622265756-ASF',
-        ],
-        'NISAR_L0B_CRSD_BETA_V1': [
-            'C1261815276-ASFDEV',
-            'C1273831262-ASF',
-            'C2850225137-ASF',
-        ],
-        'NISAR_L0B_CRSD_PROVISIONAL_V1': [
-            'C1261832632-ASFDEV',
-            'C1261832671-ASF',
-            'C2853091612-ASF',
-        ],
-        'NISAR_L0B_CRSD_V1': [
-            'C1256358463-ASFDEV',
-            'C1257349114-ASF',
-            'C3622254588-ASF',
         ],
         'NISAR_L1_RSLC_BETA_V1': [
             'C1261813489-ASFDEV',
@@ -244,11 +237,6 @@ dataset_collections = {
             'C1256729502-ASFDEV',
             'C1257349067-ASF',
             'C3622253574-ASF',
-        ],
-        'NISAR_STUF': [
-            'C1256535568-ASFDEV',
-            'C1257349089-ASF',
-            'C3622233495-ASF',
         ],
         'NISAR_OROST': [
             'C1256578011-ASFDEV',
@@ -453,6 +441,7 @@ dataset_collections = {
         'OPERA_L2_RTC-S1-STATIC_V1': ['C1259981910-ASF', 'C2795135174-ASF'],
         'OPERA_L2_RTC-S1_PROVISIONAL_V0': ['C1257995186-ASF'],
         'OPERA_L3_DISP-S1_V1': ['C3294057315-ASF', 'C1271830354-ASF'],
+        'OPERA_L4_TROPO-ZENITH_V1': ['C3717139408-ASF'],
     },
     'OPERA-S1-CALVAL': {
         'OPERA_L2_CSLC-S1_CALVAL_V1': ['C1260721945-ASF', 'C2803501758-ASF'],
@@ -613,8 +602,7 @@ dataset_collections = {
         'AIRSAR_INT': ['C1208652494-ASF'],
     },
     'SEASAT': {
-        'SEASAT_SAR_L1_TIFF': ['C1206500826-ASF', 'C1206752770-ASF'],
-        'SEASAT_SAR_L1_HDF5': ['C1206500991-ASF', 'C1206144699-ASF'],
+        'SEASAT_L1_SAR': ['C3576379529-ASF', 'C1271768606-ASF'],
     },
 }
 
@@ -957,10 +945,8 @@ collections_per_platform = {
         'C1208703384-ASF',
     ],
     'SEASAT 1': [
-        'C1206500826-ASF',
-        'C1206500991-ASF',
-        'C1206752770-ASF',
-        'C1206144699-ASF',
+        'C3576379529-ASF', # prod
+        'C1271768606-ASF', # uat
     ],
     'SMAP': [
         'C1243122884-ASF',
@@ -1081,9 +1067,6 @@ collections_per_platform = {
         'C1261815274-ASFDEV',
         'C1261832497-ASFDEV',
         'C1256358262-ASFDEV',
-        'C1261815276-ASFDEV',
-        'C1261832632-ASFDEV',
-        'C1256358463-ASFDEV',
         'C1261813489-ASFDEV',
         'C1261832868-ASFDEV',
         'C1273095154-ASFDEV',
@@ -1134,9 +1117,6 @@ collections_per_platform = {
         'C1273831320-ASF',
         'C1261832659-ASF',
         'C1257349115-ASF',
-        'C1273831262-ASF',
-        'C1261832671-ASF',
-        'C1257349114-ASF',
         'C1273831203-ASF',
         'C1261833052-ASF',
         'C1273831205-ASF',
@@ -1187,9 +1167,6 @@ collections_per_platform = {
         'C3622228339-ASF',
         'C2853089814-ASF',
         'C3622265756-ASF',
-        'C2850225137-ASF',
-        'C2853091612-ASF',
-        'C3622254588-ASF',
         'C2850225585-ASF',
         'C2853145197-ASF',
         'C3622236985-ASF',
@@ -1394,6 +1371,9 @@ collections_by_processing_level = {
         'C3294057315-ASF',
         'C1271830354-ASF'
     ],
+    'TROPO-ZENITH': [
+        'C3717139408-ASF',
+    ],
     'GRD_FD': [
         'C1214471197-ASF',
         'C1212200781-ASF',
@@ -1479,8 +1459,6 @@ collections_by_processing_level = {
         'C1207177736-ASF',
         'C1206936391-ASF',
         'C1205181982-ASF',
-        'C1206500991-ASF',
-        'C1206144699-ASF',
     ],
     '3FP': ['C1213921661-ASF', 'C1213928843-ASF', 'C1205256880-ASF', 'C1208713702-ASF'],
     'JPG': ['C1213921626-ASF', 'C1000000306-ASF'],
@@ -1492,7 +1470,6 @@ collections_by_processing_level = {
     'LSTOKES': ['C1213927939-ASF'],
     'PSTOKES': ['C1213928209-ASF'],
     'ATI': ['C1208652494-ASF'],
-    'GEOTIFF': ['C1206500826-ASF', 'C1206752770-ASF'],
     'L1A_Radar_RO_ISO_XML': [
         'C1243122884-ASF',
         'C1243141638-ASF',
@@ -1616,15 +1593,24 @@ collections_by_processing_level = {
         'C1256358262-ASFDEV', # Validated
         'C1257349115-ASF',
         'C3622265756-ASF',
-        'C1261815276-ASFDEV', # CRSD
-        'C1273831262-ASF',
-        'C2850225137-ASF',
-        'C1261832632-ASFDEV', # Provisional
-        'C1261832671-ASF',
-        'C2853091612-ASF',
-        'C1256358463-ASFDEV', # Validate
-        'C1257349114-ASF',
-        'C3622254588-ASF',
+        'C1258125097-ASFDEV',
+        'C1258836794-ASF',
+        'C2887469134-ASF',
+    ],
+    'RRSD': [
+        'C1261815274-ASFDEV', # RRSD BETA
+        'C1273831320-ASF',
+        'C3622228339-ASF',
+        'C1261832497-ASFDEV', # Provisional
+        'C1261832659-ASF',
+        'C2853089814-ASF',
+        'C1256358262-ASFDEV', # Validated
+        'C1257349115-ASF',
+        'C3622265756-ASF',
+        
+        'C1258125097-ASFDEV',
+        'C1258836794-ASF',
+        'C2887469134-ASF',
     ],
     'RSLC': [
             'C1261813489-ASFDEV', # Beta