+++ asf_search/export/metalink.py
@@ -1,6 +1,9 @@
 import inspect
+import os
 from types import GeneratorType
+from urllib import parse
 import xml.etree.ElementTree as ETree
+
 from asf_search import ASF_LOGGER
 from asf_search.export.export_translators import ASFSearchResults_to_properties_list
 
@@ -68,14 +71,32 @@ class MetalinkStreamArray(list):
 
         if p.get('md5sum') and p.get('md5sum') != 'NA':
             verification = ETree.Element('verification')
-            h = ETree.Element('hash', {'type': 'md5'})
-            h.text = p['md5sum']
-            verification.append(h)
+            if isinstance(p.get('md5sum'), dict):
+                a = parse.urlparse(p['url'])
+                file_name = os.path.basename(a.path)
+                md5_entry = p['md5sum'].get(file_name)
+                h = ETree.Element('hash', {'type': 'md5'})
+                if md5_entry is not None:
+                    h.text=md5_entry
+                    verification.append(h)
+            else:
+                h = ETree.Element('hash', {'type': 'md5'})
+                h.text = p['md5sum']
+                verification.append(h)
             file.append(verification)
 
         if p['bytes'] and p['bytes'] != 'NA':
             size = ETree.Element('size')
-            size.text = str(p['bytes'])
+            if isinstance(p.get('bytes'), dict):
+                a = parse.urlparse(p['url'])
+                file_name = os.path.basename(a.path)
+                bytes_entry = p['bytes'].get(file_name)
+                if bytes_entry is not None:
+                    size.text=str(bytes_entry['bytes'])
+                else:
+                    size.text = str(p['bytes'])
+            else:
+                size.text = str(p['bytes'])
             file.append(size)
 
         return '\n' + (8 * ' ') + ETree.tostring(file, encoding='unicode')