+++ tests/pytest-managers.py
@@ -80,6 +80,9 @@ from Serialization.test_serialization import run_test_serialization
 import nbformat
 from nbconvert.preprocessors import ExecutePreprocessor
 
+from ASFProduct.test_ASFSubproduct import run_test_ASFSubproduct
+from tests.Search.test_collection_attributes import run_test_collection_attributes
+
 
 # asf_search.ASFProduct Tests
 def test_ASFProduct(**args) -> None:
@@ -90,7 +93,19 @@ def test_ASFProduct(**args) -> None:
     geographic_response = get_resource(test_info['products'])
     run_test_ASFProduct(geographic_response)
 
+def test_ASFSubproduct(**args) -> None:
+    """
+    Tests ASFProduct subclasses for properties and basic functionality
+    """
+    session = args["config"].getoption("authenticated_session")
 
+    test_info = args['test_info']
+    scene_names = test_info['scenes']
+    expected_subclass = test_info['expected_subclass']
+    opts = ASFSearchOptions(**test_info.get('opts', {}), session=session)
+
+    run_test_ASFSubproduct(scene_names=scene_names, expected_subclass=expected_subclass, opts=opts)
+    
 def test_ASFProduct_Stack(**args) -> None:
     """
     Tests ASFProduct.stack() with reference and corresponding stack
@@ -628,6 +643,12 @@ def test_keyword_aliasing_results(**args) -> None:
 
     run_test_keyword_aliasing_results(opts)
 
+def test_collection_attributes(**args) -> None:
+    params = args['test_info']['params']
+    expected_attributes = args['test_info']['expected_attributes']
+    expect_failure = args['test_info'].get('expect_failure', False)
+    session = args["config"].getoption("authenticated_session")
+    run_test_collection_attributes(params, expected_attributes, session, expect_failure)
 
 # Finds and loads file from yml_tests/Resouces/ if loaded field ends with .yml/yaml extension
 def get_resource(yml_file):